batch: Move realized provenance storage
Realized provenance run storage now lives in realized_provenance. Realized
entry editing imports the run table, run type, claimed flag, and
line-number codecs as private collaborators.

The moved table still uses chunked mapped records with one pending Python
run. Realized entry slicing and content streaming continue to use the same
accessor path.

This commit moves provenance run merging, clipping, lookup, and
mapped-record storage out of realized_entries.

The realized entry module now focuses on editor-backed content operations
and entry views while provenance storage has its own module boundary.
